+++ Blink/SmartKeys/SmartKeysController.m
@@ -155,7 +155,7 @@ static NSArray *CursorKeys = nil;
   [self.allKeys addObject:[[SmartKey alloc] initWithName:KbdEscKey symbol:UIKeyInputEscape]];
 }
 
-- (void)symbolUp:(NSString *)symbol
+- (void)invalidateTimer
 {
   if (_timer != nil) {
     [_timer invalidate];
@@ -163,8 +163,16 @@ static NSArray *CursorKeys = nil;
   }
 }
 
+- (void)symbolUp:(NSString *)symbol
+{
+  [self invalidateTimer];
+}
+
 - (void)symbolDown:(NSString *)symbol
 {
+  // Always invalidate the previous key press if there was one.
+  [self invalidateTimer];
+  
   for (SmartKey *key in self.allKeys) {
     if ([key.name isEqualToString:symbol]) {
       [_textInputDelegate insertText:key.symbol];